react reorganization
