+++ front-end/package.json
@@ -10,7 +10,8 @@
     "react-router": "^4.3.1",
     "react-router-dom": "^4.3.1",
     "react-scripts": "2.1.3",
-    "react-stripe-elements": "^2.0.1"
+    "react-stripe-elements": "^2.0.1",
+    "styled-components": "^4.1.3"
   },
   "scripts": {
     "start": "react-scripts start",
