+++ front-end/src/App.js
@@ -1,14 +1,18 @@
 import React, { Component } from "react";
+import { BrowserRouter as Router } from "react-router-dom";
 import "./App.css";
 import Routes from "./components/Routes/Routes.js";
 
 class App extends Component {
   render() {
     return (
-      <div className="App">
-        <Routes />
-        {/* <Footer /> */}
-      </div>
+      <Router>
+        <div className="App">
+          {/* <NavBar /> */}
+          <Routes />
+          {/* <Footer /> */}
+        </div>
+      </Router>
     );
   }
 }
