+++ front-end/src/components/Landing/job.js
