+++ front-end/src/index.js
@@ -2,15 +2,12 @@ import React from "react";
 import ReactDOM from "react-dom";
 import "./index.css";
 import App from "./App";
-import { BrowserRouter as Router } from "react-router-dom";
 
 import Firebase, { FirebaseContext } from "./components/Firebase";
 
 ReactDOM.render(
   <FirebaseContext.Provider value={new Firebase()}>
-    <Router>
-      <App />
-    </Router>
+    <App />
   </FirebaseContext.Provider>,
   document.getElementById("root")
 );
